Refactor following feedback
