+++ Week3/routes/tag.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const { getTodobyTag } = require('../helpers');
+const { getTodobyTag } = require('../handlers/tag');
 
 //* Get all Todos by tag
 router.get('/:tagname', getTodobyTag);
